fix: update clap and reorg into arg structs
This makes it more readable and allows for better reuse. And defaults! CLI was never this good!
