+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod app;
 mod embeded;
+pub mod args;
